Move code to EE folder, and add EE license
