+++ LICENSE.md
@@ -1,3 +1,12 @@
+The code in the root directory and all subdirectories, except for the 'ee' folder,
+is licensed under the Mozilla Public License Version 2.0 (the "MPL"), as detailed below.
+
+The code in the 'ee' folder is subject to a separate commercial license.
+See the [LICENSE-EE](ee/LICENSE-EE) file in the 'ee' folder for details on the Enterprise Edition license.
+
+For the open source components:
+-------------------------------
+
 # Mozilla Public License Version 2.0
 
 ## 1. Definitions
