--- ee/LICENSE-EE
@@ -1,7 +1,0 @@
-Commercial License
-
-The code in this 'ee' folder is subject to the commercial license terms outlined in the accompanying EULA.pdf file.
-
-For the full End User License Agreement, please refer to the EULA.pdf file in this directory.
-
-Copyright 2024 Tyk Technologies. All rights reserved.
