+++ ee/LICENSE-EE.md
@@ -1,0 +1,7 @@
+Commercial License
+
+The code in this 'ee' folder is subject to the commercial license terms outlined in the accompanying [EULA.pdf](ee/EULA.pdf) file.
+
+For the full End User License Agreement, please refer to the [EULA.pdf](ee/EULA.pdf) file in this directory.
+
+Copyright 2024 Tyk Technologies. All rights reserved.
