+++ ee/internal/middleware/streams/middleware.go
@@ -1,4 +1,4 @@
-package streamv1
+package streams
 
 import (
 	"context"
