+++ ee/internal/middleware/streams/model.go
@@ -1,4 +1,4 @@
-package streamv1
+package streams
 
 import (
 	"sync/atomic"
